Refactoring component names
